Swapped the data type for the stop coordinates to double

+++ database/migrations/2019_04_12_072444_create_stops.php
@@ -18,8 +18,8 @@ class CreateStops extends Migration
             $table->string('name');
             $table->string('short_desc');
             $table->string('full_desc');
-            $table->decimal('coord_x', 10, 6);
-            $table->decimal('coord_y', 10, 6);
+            $table->double('coord_x', 10, 6);
+            $table->double('coord_y', 10, 6);
             $table->timestamps();
         });
     }
